+++ src/constants/media.ts
@@ -1,5 +1,5 @@
-export const DEFAULT_AVATAR = require('../../assets/images/default-avatar.jpeg');
+export const DEFAULT_AVATAR = require('../../assets/images/avatar.png');
 export const DEFAULT_IMAGE = require('../../assets/images/backdrop.png');
 
-export const DEFAULT_AVATAR_DIMENSIONS = { width: 200, height: 200 };
+export const DEFAULT_AVATAR_DIMENSIONS = { width: 256, height: 256 };
 export const DEFAULT_IMAGE_DIMENSIONS = { width: 600, height: 400 };